Use sudo to restart unicorn
We probably have to do this now due to running almost the whole installer script with escalating privileges[1][2].

We should review the Installer and only escalate privileges when strictly necessary.

In the meantime, this commit should help avoid a file permission problem when restarting the application with `cap production deploy:restart`.

[1] https://github.com/consul/installer/pull/97/commits/f406a2f1ceb330f065360c20ece9008ebaf9c9e4

[2] https://docs.ansible.com/ansible/latest/user_guide/become.html